requirements.txt
fastapi
uvicorn
pandas
python-multipart
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
sentence-transformers